% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseImpact.R
\name{baseImpact}
\alias{baseImpact}
\title{Calculate base variables impact (function for users)}
\usage{
baseImpact(paths, weights = NULL)
}
\arguments{
\item{paths}{- paths to files with data}

\item{weights}{- weights for importance of variables in each model (default: NULL, equal)}
}
\value{
dataframe with columns: variable (base variable), impact (calculated based on several models specified in mdoel_ids), model_id
}
\description{
Calculate base variables impact (function for users)
}
