% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseImpact_combined.R
\name{baseImpact_combined}
\alias{baseImpact_combined}
\title{Calculate base variables importance for combined model (function for users)}
\usage{
baseImpact_combined(
  paths,
  model_ids,
  ensamble_model_id,
  weights = NULL,
  divide_by_number = FALSE
)
}
\arguments{
\item{paths}{- paths to files with data}

\item{model_ids}{- id of component models}

\item{weights}{- weights for importance of variables in each model (default: NULL, equal)}

\item{divide_by_number}{- boolean, do you want to divide impact by number of base variables}

\item{ensamble_model_ids}{- combined model id}
}
\value{
dataframe with columns: variable (base variable), impact (calculated based on several models specified in mdoel_ids), model_id
}
\description{
Calculate base variables importance for combined model (function for users)
}
