% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_combined_impact.R
\name{calculate_combined_model}
\alias{calculate_combined_model}
\title{Calculate combined model basic variables impact}
\usage{
calculate_combined_model(
  importance_frame,
  model_ids,
  ensamble_model_id,
  weights = NULL
)
}
\arguments{
\item{importance_frame}{- dataframe with columns: variable (base variable), impact, model_id}

\item{model_ids}{- id of component models}

\item{ensamble_model_id}{- id for combined model}

\item{weights}{- weights for importance of variables in each model (default: NULL, equal)}
}
\value{
res - dataframe with columns: variable (base variable), impact (calculated based on several models specified in mdoel_ids), model_id
}
\description{
Calculate combined model basic variables impact
}
