% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_imact_based_variables.R
\name{calculate_impact}
\alias{calculate_impact}
\title{Sum up impacts of the variable including weights}
\usage{
calculate_impact(base_variables_importance, metric_cols, weights = NULL)
}
\arguments{
\item{base_variables_importance}{- dataframe with calculated base variables importance}

\item{metric_cols}{- column names with metrics}

\item{weights}{- weights for importance of variables in each model (default: NULL, equal)}
}
\value{
temp_frame - dataframe with the combined impact of the variables
}
\description{
Sum up impacts of the variable including weights
}
