% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_variables.R
\name{classify_variables}
\alias{classify_variables}
\title{Calculate class for each base variable}
\usage{
classify_variables(frame, base_columns, metric_columns)
}
\arguments{
\item{frame}{- dataframe with columns: model variable name, base variable name, model id, and the rest of the columns concern the impact of the variables}

\item{base_columns}{- model variable name, base variable name, model id column names}

\item{metric_columns}{- matric column names}
}
\value{
dataframe with columns: model variable name, base variable name, model id, class, and the rest of the columns concern the impact of the variables
}
\description{
Calculate class for each base variable
}
