% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decreasing_frequency_importance.R
\name{decreasing_frequency_importance}
\alias{decreasing_frequency_importance}
\title{Calculate base variables importance for each model}
\usage{
decreasing_frequency_importance(frame, base_columns, divide_by_number)
}
\arguments{
\item{frame}{- dataframe with columns: model variable name, base variable name, model id, class, and the rest of the columns concern the impact of the variables}

\item{base_columns}{- names for first four columns (default: c("model_variable", "base_variable", "model_id", "class"))}

\item{divide_by_number}{- boolean, do you want to divide impact by number of base variables}
}
\value{
dataframe with columns: variable (base variable), metric, decreasing (calculated impact)
}
\description{
Calculate base variables importance for each model
}
