% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decreasing_importance.R
\name{decreasing_importance}
\alias{decreasing_importance}
\title{Calculate decreasing importance}
\usage{
decreasing_importance(base_imp, rest_imp)
}
\arguments{
\item{base_imp}{- highest importance for specified variable}

\item{rest_imp}{- the rest of importances for specified variable}
}
\value{
sum_imp - numeric value with base variable importance
}
\description{
Calculate decreasing importance
}
