% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Opposite to %in% function}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{- vector with values that are in y}

\item{y}{- vector to verify weather there are values from x in it}
}
\value{
vector with values that are in x and not in y
}
\description{
Opposite to %in% function
}
