% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_impact_class.R
\name{push_impact_class}
\alias{push_impact_class}
\title{Calculate base variables importance for combined model (function for users)}
\usage{
push_impact_class(connection, impact)
}
\arguments{
\item{connection}{- connection to database}

\item{impact}{- data frame with columns: variable name, impact, class, model_id (baseImpact_combined ora baseImpact result)}
}
\description{
Calculate base variables importance for combined model (function for users)
}
