% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_quantiles.R
\name{simple_quantiles}
\alias{simple_quantiles}
\title{Calculate class based on specific metric values}
\usage{
simple_quantiles(ms)
}
\arguments{
\item{ms}{- specific metric values (for example Information.gain)}
}
\value{
vector with classes
}
\description{
Calculate class based on specific metric values
}
