% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseImpact_combined.R
\name{baseImpact_combined}
\alias{baseImpact_combined}
\title{Calculate base variables importance for combined model (function for users)}
\usage{
baseImpact_combined(
  paths,
  model_ids,
  ensamble_model_id,
  weights_metrics = NULL,
  weights_models = NULL
)
}
\arguments{
\item{paths}{- paths to files with data}

\item{model_ids}{- id of component models (vector of numeric values)}

\item{weights_metrics}{- weights for importance of variables in each model (default: NULL, equal)}

\item{weights_models}{- weights for importance of models impact (default: NULL, equal)}

\item{ensamble_model_ids}{- combined model id}
}
\value{
dataframe with columns: variable (base variable), impact (calculated based on several models specified in mdoel_ids), model_id
}
\description{
Calculate base variables importance for combined model (function for users)
}
